requirements-vercel.txt
flask>=2.3.0
openai>=1.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
pillow-heif>=1.0.0
pandas>=2.0.0
werkzeug>=2.3.0
plaid-python>=9.0.0
gunicorn>=20.1.0
